dynamic mcp: simplify context toolrunner cleanup
